Rainbow Chaotic Madness
I've also showed it to anthony and I think he liked it a lot in the
email! I love the 3-D camera you added though. Maybe we can incorporate
that into the sketch? I’ll try that now myself. The camera was cool but
the drawing as a whole was a bit bland and I’d like to incorporate
camera into the rainbow madness.